Authorization almost done

+++ app/Http/Controllers/ProjectController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\Project\StoreProjectRequest;
 use App\Http\Requests\Project\UpdateProjectRequest;
+use App\Models\Project;
 use App\Services\ProjectService;
-
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
@@ -13,8 +14,10 @@ class ProjectController extends Controller
     {
     }
 
-    public function index()
+    public function index(Project $project)
     {
+        $this->authorize('viewProject', $project);
+
         return view('pages.projects.index');
     }
 
@@ -23,8 +26,10 @@ class ProjectController extends Controller
         return $this->projectService->yajraDataTable();
     }
 
-    public function store(StoreProjectRequest $request)
+    public function store(StoreProjectRequest $request, Project $project)
     {
+        $this->authorize('createProject', $project);
+
         $result = $this->projectService->createProject($request);
 
         return response()->json($result['alertMsg'], $result['statusCode']);
@@ -37,15 +42,19 @@ class ProjectController extends Controller
         return response()->json($result);
     }
 
-    public function update(UpdateProjectRequest $request, $projectId)
+    public function update(UpdateProjectRequest $request, $projectId, Project $project)
     {
+        $this->authorize('updateProject', $project);
+
         $result = $this->projectService->updateProject($request, $projectId);
 
         return response()->json($result['alertMsg'], $result['statusCode']);
     }
 
-    public function destroy($projectId)
+    public function destroy($projectId, Project $project)
     {
+        $this->authorize('deleteProject', $project);
+
         $result = $this->projectService->removeProject($projectId);
 
         return response()->json($result['alertMsg'], $result['statusCode']);
